1.4.3

See merge request f-matano44/pysaten!30

+++ src/pysaten/utility/color_noise.py
@@ -7,7 +7,7 @@ def white(length: int, seed: int, device: str = "cpu") -> torch.Tensor:
     return torch.rand(length, generator=gen, device=device) * 2.0 - 1.0
 
 
-def blue(length: int, sr: int, seed: int, device: str = "cpu") -> torch.Tensor:
+def blue(length: int, sr: float, seed: int, device: str = "cpu") -> torch.Tensor:
     offset = int(length / 2)
     # white noise
     wh = white(length + (offset * 2), seed, device)
@@ -24,7 +24,7 @@ def blue(length: int, sr: int, seed: int, device: str = "cpu") -> torch.Tensor:
     return bl[offset : length + offset]
 
 
-def pink(length: int, sr: int, seed: int, device: str = "cpu") -> torch.Tensor:
+def pink(length: int, sr: float, seed: int, device: str = "cpu") -> torch.Tensor:
     offset = int(length / 2)
     # white noise
     wh = white(length + (offset * 2), seed, device)
